refactor: extract quick filter empty state into a separate component
